+++ engine/common/soundlib/snd_main.c
@@ -59,7 +59,7 @@ wavdata_t *FS_LoadSound( const char *filename, const byte *buffer, size_t size )
 	const char	*ext = COM_FileExtension( filename );
 	string		path, loadname;
 	qboolean		anyformat = true;
-	int		filesize = 0;
+	size_t		filesize = 0;
 	const loadwavfmt_t	*format;
 	byte		*f;
 
@@ -223,7 +223,7 @@ FS_ReadStream
 extract stream as wav-data and put into buffer, move file pointer
 ================
 */
-long FS_ReadStream( stream_t *stream, int bytes, void *buffer )
+int FS_ReadStream( stream_t *stream, int bytes, void *buffer )
 {
 	if( !stream || !stream->format || !stream->format->readfunc )
 		return 0;
@@ -241,7 +241,7 @@ FS_GetStreamPos
 get stream position (in bytes)
 ================
 */
-long FS_GetStreamPos( stream_t *stream )
+int FS_GetStreamPos( stream_t *stream )
 {
 	if( !stream || !stream->format || !stream->format->getposfunc )
 		return -1;
@@ -256,7 +256,7 @@ FS_SetStreamPos
 set stream position (in bytes)
 ================
 */
-long FS_SetStreamPos( stream_t *stream, long newpos )
+int FS_SetStreamPos( stream_t *stream, int newpos )
 {
 	if( !stream || !stream->format || !stream->format->setposfunc )
 		return -1;
